fix(api): views_count salía siempre 0 y pages_count no aparecía nunca
Sale de revisar si la documentación seguía siendo cierta después de los
arreglos. `docs/info/api/v2/content.md` describía tres campos rotos como si
fueran el comportamiento normal — uno ya lo había arreglado, los otros dos no.

views_count
-----------
El resource expone `$this->views_count`, pero ninguna consulta lo agregaba:
salía 0 para todo el mundo. Y las visitas SÍ se cuentan — `ProcessContentViewJob`
las escribe en `content_daily_views` en cada petición de un contenido. O sea
que se estaba contabilizando el dato y no se devolvía nunca.

Ahora se agrega con `withSum('dailyViews as views_count', 'views')`. Se
castea a int porque `withSum` devuelve null, no 0, cuando aún no hay visitas.

pages_count
-----------
Depende de `whenCounted('pages')` y nadie hacía el `withCount`, así que la
clave no aparecía en ninguna respuesta. Se añade el conteo.

Los dos van como agregados de la consulta que ya se hacía, así que no añaden
una consulta por elemento en los listados.

Documentación
-------------
`content.md` decía que estos campos salían siempre a null/0/ausentes. Ahora
describe lo que devuelven de verdad, y deja anotado lo que hacían antes para
que se entienda un cambio de respuesta si alguien lo nota desde una web.

+++ app/Http/Resources/V2/Content/ContentResource.php
@@ -59,7 +59,9 @@ class ContentResource extends JsonResource
             'tags' => $this->whenLoaded('tags'),
             'technologies' => $this->whenLoaded('technologies'),
             'pages_count' => $this->whenCounted('pages'),
-            'views_count' => $this->views_count ?? 0,
+            // `withSum` devuelve null cuando el contenido no tiene ninguna
+            // visita todavía, no 0.
+            'views_count' => (int) ($this->views_count ?? 0),
             'published_at' => $this->published_at?->toISOString(),
             'created_at' => $this->created_at?->toISOString(),
             'updated_at' => $this->updated_at?->toISOString(),
